got a test for highlighting
Using the built in espuds face tests didn't seem to work with the
propertizing functions I'm using in the goalie code. I'm guessing its
not part of the "face" but an extra property. So checking directly
against those properties seemed to work. (This took a whole day to
figure out so remember it when you are old and wonder where your life went)